+++ client/src/main/java/net/eanfang/client/ui/widget/CompanyListView.java
@@ -2,8 +2,6 @@ package net.eanfang.client.ui.widget;
 
 import android.app.Activity;
 import android.graphics.drawable.ColorDrawable;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.WindowManager;
@@ -26,6 +24,9 @@ import net.eanfang.client.R;
 import net.eanfang.client.ui.adapter.SwitchCompanyListAdapter;
 
 import java.util.List;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Created by MrHou
@@ -55,8 +56,8 @@ public class CompanyListView extends PopupWindow {
         adapter.bindToRecyclerView(revCompanyList);
         revCompanyList.setLayoutManager(new LinearLayoutManager(mContext));
         setBackgroundDrawable(new ColorDrawable(0x70000000));
-        int width = (int) (ScreenUtils.widthPixels(context));
-        int height = (int) (ScreenUtils.heightPixels(context) * 3 / 4);
+        int width = (ScreenUtils.widthPixels(context));
+        int height = (ScreenUtils.heightPixels(context) * 3 / 4);
         setWidth(width);
         setHeight(WindowManager.LayoutParams.WRAP_CONTENT);
         setFocusable(true);
